ADF write support & FAT fixes WIP

+++ include/os_bito.h
@@ -4,22 +4,22 @@
 
 static inline uint16_t read16LE(const uint8_t* ptr)
 {
-	return ptr[0] | (ptr[1] << 8);
+	return (uint16_t)ptr[0] | ((uint16_t)ptr[1] << 8);
 }
 
 static inline uint32_t read32LE(const uint8_t* ptr)
 {
-	return ptr[0] | (ptr[1] << 8) | (ptr[2] << 16) | (ptr[3] << 24);
+	return (uint32_t)ptr[0] | ((uint32_t)ptr[1] << 8) | ((uint32_t)ptr[2] << 16) | ((uint32_t)ptr[3] << 24);
 }
 
 static inline uint16_t read16BE(const uint8_t* ptr)
 {
-	return ptr[1] | (ptr[0] << 8);
+	return (uint16_t)ptr[1] | ((uint16_t)ptr[0] << 8);
 }
 
 static inline uint32_t read32BE(const uint8_t* ptr)
 {
-	return ptr[3] | (ptr[2] << 8) | (ptr[1] << 16) | (ptr[0] << 24);
+	return (uint32_t)ptr[3] | ((uint32_t)ptr[2] << 8) | ((uint32_t)ptr[1] << 16) | ((uint32_t)ptr[0] << 24);
 }
 
 static inline void write16(uint8_t* ptr, uint16_t n, bool littleEndian)
@@ -57,17 +57,17 @@ static inline void write32(uint8_t* ptr, uint32_t n, bool littleEndian)
 static inline uint16_t read16(const uint8_t* ptr, bool littleEndian)
 {
 	if (littleEndian)
-		return ptr[0] | (ptr[1] << 8);
+		return (uint16_t)ptr[0] | ((uint16_t)ptr[1] << 8);
 	else
-		return ptr[1] | (ptr[0] << 8);
+		return (uint16_t)ptr[1] | ((uint16_t)ptr[0] << 8);
 }
 
 static inline uint32_t read32(const uint8_t* ptr, bool littleEndian)
 {
 	if (littleEndian)
-		return ptr[0] | (ptr[1] << 8) | (ptr[2] << 16) | (ptr[3] << 24);
+		return (uint32_t)ptr[0] | ((uint32_t)ptr[1] << 8) | ((uint32_t)ptr[2] << 16) | ((uint32_t)ptr[3] << 24);
 	else
-		return ptr[3] | (ptr[2] << 8) | (ptr[1] << 16) | (ptr[0] << 24);
+		return (uint32_t)ptr[3] | ((uint32_t)ptr[2] << 8) | ((uint32_t)ptr[1] << 16) | ((uint32_t)ptr[0] << 24);
 }
 
 static inline uint16_t fix16(uint16_t value, bool littleEndian)
